refractor: ProductControllerTest 분해
- 단일 테스트에 너무 많은 테스트 케이스를 수행하고 있어 ProductControllerTest를 기능 단위로 분해했습니다.
