+++ src/test/java/gift/AbstractControllerTest.java
@@ -1,8 +1,13 @@
 package gift;
 
+import gift.dto.ErrorMessageResponse;
+import org.junit.jupiter.api.Assertions;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestClient;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public abstract class AbstractControllerTest {
@@ -14,4 +19,13 @@ public abstract class AbstractControllerTest {
     protected String getBaseUrl() {
         return "http://localhost:" + port;
     }
+
+    protected void checkErrorResponse(HttpClientErrorException e, int expectedStatus) {
+        assertThat(e.getStatusCode().value()).isEqualTo(expectedStatus);
+        ErrorMessageResponse errorResponse = e.getResponseBodyAs(ErrorMessageResponse.class);
+        assertThat(errorResponse).isNotNull();
+        Assertions.assertNotNull(errorResponse);
+        Assertions.assertEquals(expectedStatus, errorResponse.status(), "상태 코드는 " + expectedStatus + "이어야 합니다.");
+    }
+
 }
